+++ src/components/InstallAppButton.module.css
@@ -1,6 +1,4 @@
 .install_button {
   margin-top: 0.3rem;
   font-weight: 500;
-  /* background-color: #791fbd;
-  color: white; */
 }
