+++ src/components/auth/LogoutButton.tsx
@@ -1,4 +1,5 @@
 import styles from './LogoutButton.module.css';
+
 const Logout = () => {
   return (
     <form className={styles.logout} action="/auth/signout" method="post" title="잘가올...">
