+++ src/hooks/useQueryUser.ts
@@ -1,7 +1,7 @@
 import { getSession } from '@/api/auth/getSession';
 import { createClient } from '@/utils/supabase/client';
-import type { User } from '@supabase/supabase-js';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
+import type { User } from '@supabase/supabase-js';
 
 export const useQueryUser = () => {
   const supabase = createClient();
